pyproject.toml
[project]
name = "md2pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.120.0",
    "jinja2>=3.1.6",
    "markdown>=3.9",
    "pygments>=2.19.2",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
]
